requirements.txt
beautifulsoup4
selenium>=2.44.0,<3.0.0
requests
unidecode
vcrpy
future
fake-useragent
telepot
Google-Search-API
boto3
awscli



